update unit and cypress tests [skip ci pytest]
