+++ functionnal_tests/cypress/integration/sidebar/presence_sidebar_allcontent.js
@@ -7,13 +7,13 @@ describe('Sidebar', function () {
   beforeEach(function () {
     cy.loginAs('users')
     cy.visit('/ui')
-    cy.get('.sidebar__content__navigation__workspace__item__number').click()
+    cy.get('.sidebar__content__navigation__workspace__item__name').click()
   })
   it('should have a link to All contents in the hidden menu', function () {
     cy.get('.sidebar__content__navigation__workspace__item__menu').should('be.visible').click()
     cy.get('li').contains('All Contents').should('have.attr', 'href', '/ui/workspaces/1/contents')
-    cy.get('.fa-th').should('be.visible')
-    cy.get('.fa-th').click()
+    cy.get('[data-cy="sidebar_subdropdown-contents/all"]').should('be.visible')
+    cy.get('[data-cy="sidebar_subdropdown-contents/all"]').click()
     cy.url().should('include', '/workspaces/1/contents')
     cy.get('.workspace__header.pageTitleGeneric').should('be.visible')
   })
