+++ functionnal_tests/cypress/integration/sidebar/presence_sidebar_dashboard.js
@@ -11,12 +11,12 @@ describe('Sidebar', function () {
     cy.visitPage({
       pageName: p.HOME
     })
-    cy.get('.sidebar__content__navigation__workspace__item__number').click()
+    cy.get('.sidebar__content__navigation__workspace__item__name').click()
   })
   it('should have a link to Dashboard in the hidden menu', function () {
     cy.get('.sidebar__content__navigation__workspace__item__menu').should('be.visible').click()
     cy.get('li').contains('Dashboard').should('have.attr', 'href', '/ui/workspaces/1/dashboard')
-    cy.get('.fa-home').should('be.visible').click()
+    cy.get('[data-cy="sidebar_subdropdown-dashboard"]').should('be.visible').click()
     cy.url().should('include', '/workspaces/1/dashboard')
     cy.get('.dashboard__header.pageTitleGeneric').should('be.visible')
   })
